• Category is now a sub property of SFSymbol
