+++ Sources/SFSymbol/SFSymbol/Multicolor.swift
