+++ Sources/SFSymbol/SFSymbol/Restricted+SFSymbol.swift
