+++ Sources/SFSymbol/SFSymbol2/Image+SFSymbol2.swift
