+++ Sources/SFSymbol/SFSymbol2/UIImage+SFSymbol2.swift
